requirements.txt
streamlit>=1.28.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
pathlib2>=2.3.7
typing-extensions>=4.7.0
